refactor(next): correct typo in storybook generator
