+++ .gitignore
@@ -1,4 +1,5 @@
 # See http://help.github.com/ignore-files/ for more about ignoring files.
+migrations.json
 
 # compiled output
 dist
